Coding Days: warnings fixes grid/generation
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2666 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/generation/PrewaveletGridGenerator.cpp
@@ -70,13 +70,13 @@ void PrewaveletGridGenerator::full(size_t level)
 void PrewaveletGridGenerator::refine(RefinementFunctor* func)
 {
 	HashRefinement refine;
-	int start = this->storage->size();
+	size_t start = this->storage->size();
 	refine.free_refine(this->storage, func);
-	int end = this->storage->size();
+	size_t end = this->storage->size();
 	//All added gridpoint are between [start,end[
 
 	//Check if a gridpoint within the shadow storage is now part of the actual grid!
-	for (int i = start; i < end; i++)
+	for (size_t i = start; i < end; i++)
 	{
 		if (shadowstorage->find(storage->get(i)) != shadowstorage->end())
 		{
@@ -86,7 +86,7 @@ void PrewaveletGridGenerator::refine(RefinementFunctor* func)
 	}
 
 	//Now add all missing neigbours to the shadowStorage
-	for (int i = start; i < end; i++)
+	for (size_t i = start; i < end; i++)
 	{
 		GridStorage::index_pointer index = this->storage->get(i);
 
@@ -201,13 +201,13 @@ void PrewaveletGridGenerator::addNeighbours(index_type& index,
 
 				// The index cast to int is required to allow a negative index
 				int target_left = (1.0 / (1 << target_level))
-						* ((int) (target_index) - 3);
+						* static_cast<double> (target_index - 3);
 				int target_right = (1.0 / (1 << target_level))
-						* ((int) (target_index) + 3);
+						* static_cast<double> (target_index + 3);
 				int current_left = (1.0 / (1 << current_index))
-						* ((int) (current_level) + 3);
+						* static_cast<double> (current_level + 3);
 				int current_right = (1.0 / (1 << current_index))
-						* ((int) (current_level) + 3);
+						* static_cast<double> (current_level + 3);
 
 				if (!(current_right > target_left || current_left
 						< target_right))
